+++ lib/src/settings.dart
@@ -11,6 +11,7 @@ class SettingsList extends StatelessWidget {
     super.key,
     required this.sections,
     this.clipBehavior = Clip.hardEdge,
+    this.controller,
     this.padding,
     this.physics,
     this.shrinkWrap = false,
@@ -22,6 +23,9 @@ class SettingsList extends StatelessWidget {
 
   /// The content will be clipped (or not) according to this option.
   final Clip clipBehavior;
+
+  /// An optional controller for the scroll view.
+  final ScrollController? controller;
 
   /// The amount of space to surround the child.
   final EdgeInsets? padding;
@@ -37,6 +41,7 @@ class SettingsList extends StatelessWidget {
   Widget build(BuildContext context) {
     return ListView.builder(
       clipBehavior: clipBehavior,
+      controller: controller,
       padding: padding,
       physics: physics,
       shrinkWrap: shrinkWrap,